requirements.txt
pycryptodome
flask
flask-socketio
eventlet